requirements.txt
openai>=1.3.0
huggingface_hub>=0.19.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.13.0
accelerate>=0.20.0
